pyproject.toml
[project]
name = "agent-with-mcp"
version = "0.1.0"
description = "Conversational AI agent with MCP tools, Ollama, and terminal UI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
keywords = ["agent", "mcp", "ollama", "langgraph", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langchain>=1.2.10",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.4.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langmem>=0.0.20",
    "rich>=14.0.0",
    "prompt-toolkit>=3.0.50",
    "python-dotenv>=1.1.0",
    "psycopg[binary]>=3.2.0",
]

[project.scripts]
agent-with-mcp = "main:main"
